+++ nba_player_news/data/publishers.py
@@ -16,7 +16,7 @@ from environment import REDIS_URL, REDIS_SUBSCRIPTION_MESSAGES_CHANNEL_NAME
 from environment import TWITTER_ACCESS_SECRET, TWITTER_ACCESS_TOKEN, TWITTER_CONSUMER_KEY, TWITTER_CONSUMER_SECRET
 from nba_player_news.data.messages import SubscriptionMessage
 from nba_player_news.data.senders import Emailer, Tweeter
-from nba_player_news.data.sent_message_builders import FacebookMessengerMessageBuilder
+from nba_player_news.data.message_builders import FacebookMessengerMessageBuilder
 from nba_player_news.models import Subscription
